+++ src/main/java/ubco/cosc520/graph/TwoGraphUnionEdgeCount.java
@@ -6,7 +6,7 @@ import org.apache.commons.math3.linear.RealMatrix;
  * Calculates the untion between two graphs.
  */
 public class TwoGraphUnionEdgeCount
-        implements TwoGraphOperator<Integer> {
+        implements TwoGraphOperator<Double> {
 
     /**
      * Calculates the union between two graphs.
@@ -16,7 +16,7 @@ public class TwoGraphUnionEdgeCount
      * @return A new graph, which is the union of G and H.
      */
     @Override
-    public Integer operate(
+    public Double operate(
             final Graph g,
             final Graph h
     ) {
@@ -26,7 +26,7 @@ public class TwoGraphUnionEdgeCount
         int size = rmg.getColumn(0).length;
 
         //calculate total number of edges for union graph of g and h
-        int edgeNumOfUnion = 0;
+        double edgeNumOfUnion = 0;
         for (int i = 0; i < size; i++) {
             for (int j = 0; j < size; j++) {
                 //skip dignose line
